Optimize algorithm by sorting in-place
This finishes the optimizations by relying on the fact that we're
passing a reference to the list which means we can modify it in-place.
We sort the whole list of symbols on the first pass only, and then as
sort children into parents' buckets we *move* them by adding them to
`parent.Children` and removing them from the current list, and then
recurse and modify those lists until we're done.

I don't normally like relying on side-effects such as modifying in the
called function, but it makes the most sense in this case for
optimizations and is duly noted.