Allow `WorkspacePaths` to be empty if we're not in a workspace
We were previously adding the initial working directory as a fallback if
there was no workspace. However, this led to new VS Code windows (not in
a workspace) enumerating all files in the home directory (if that's what
the client had to default to, without a workspace folder). And this
could even spam permission requests on more secure systems such as
macOS. So we just need to let be an empty enumerable.