When the built-in `$null` was watched its value was incorrect
This fixes the evaluation of `$null` itself and variables that don't
exist such that they display correctly in the watched variables pane of
the debugger. As noted, the issue was that we were doing a string join
even for a single (potentially null) value.

This also fixes the excess quotes bug around a single value.