Update shell integration script to fix command decorations
Happy new year! Updated one year ago today, it needed it again. This has
actually been broken for a while unfortunately because the last time we
manually patched it to disable strict mode, we did so before capturing
the exit code (and thus it was lost). Fixed now at least.