WIP: Evaluate variables on hover
If their value is available (either because its been set in the session
manually, or the script has been executed or is being debugged) we can
return it on the hover.

We should ask if we actually want to do this, and if so, how we want to
format it.