Add sane defaults to `Start-EditorServices` and fix logs
We do not need to require so many CLI flags.

This continues to run the existing Emacs and Vim tests utilizing those
flags as a regression scenario, and adds an additional pair of tests
that launch PSES with a minimal set of flags.

This also fixes the log file path to be unique per process, and to use `LogPath`
as a directory instead of a file.