Remove `PowerShellEditorServices.VSCode` module / `ContentViews` feature
As it took a year (presumably, maybe a more recent VS Code update is
what broke it) to come up that indicates its usage is too low for us to
go about fixing. If someone wants to figure it out we'll be happy to
take a PR. For now we're removing the code (which is also a massive
simplification because it was the entirety of the "VSCode" module we had
to build and ship with the extension).