Use a provided shell integration script directly
Instead of having to maintain an edited copy (which was really annoying)
I stubbed out `PSConsoleHostReadLine` to do what's expected.
So now we can just use the existing shell integration script directly!

Since we can't reliably find the script using
`code --locate-shell-integration-path pwsh` we now rely on it being sent
by the client on initialization. Its presence implies the feature is on.

This is pretty VS Code specific, but not necessarily so.

Apply suggestions from code review

Thanks Patrick!

Co-authored-by: Patrick Meinecke <SeeminglyScience@users.noreply.github.com>