Switch OneBranch pipeline to only build and not test
Primarily this is because the end-to-end tests currently hang only on
OneBranch because the image still only has end-of-life PowerShell 7.3,
which we've dropped support for. But we need to do a release.

All the CI tests run on GitHub where we do our actual development and
their images are up-to-date so it all passes.